code/device-data-service/pyproject.toml
[tool.poetry]
name = "device-data-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["all"], version = "^0.105.0" }
pydantic-settings = "^2.1.0"
psycopg = { extras = ["binary"], version = "^3.1.16" }
psycopg-pool = "^3.2.0"
loguru = "^0.7.2"
pyjwt = "^2.8.0"
aiohttp = "^3.9.1"
cachetools = "^5.3.2"
aiokafka = "^0.10.0"
orjson = "^3.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
